Fix issues from #65: add TypeScript types, extract shared validation, fix async interface for window operations
- Set noImplicitAny to true in tsconfig.json
- Created shared validation module to eliminate duplication
- Fixed async interface workaround in screen.ts by properly making methods async
- Updated tests to use await with async methods
