+++ src/providers/keysender/keyboard.ts
@@ -6,84 +6,18 @@ type KeyboardButtonType = string;
 import { KeyboardInput, KeyCombination, KeyHoldOperation } from '../../types/common.js';
 import { WindowsControlResponse } from '../../types/responses.js';
 import { KeyboardAutomation } from '../../interfaces/automation.js';
-
-// Constants for validation
-const MAX_TEXT_LENGTH = 1000;
-
-// List of allowed keyboard keys for validation
-const VALID_KEYS = [
-  // Letters
-  'a', 'b', 'c', 'd', 'e', 'f', 'g', 'h', 'i', 'j', 'k', 'l', 'm',
-  'n', 'o', 'p', 'q', 'r', 's', 't', 'u', 'v', 'w', 'x', 'y', 'z',
-  // Numbers
-  '0', '1', '2', '3', '4', '5', '6', '7', '8', '9',
-  // Special keys
-  'space', 'escape', 'tab', 'alt', 'control', 'shift', 'right_shift',
-  'command', 'enter', 'return', 'backspace', 'delete', 'home', 'end',
-  'page_up', 'page_down', 'left', 'up', 'right', 'down',
-  // Function keys
-  'f1', 'f2', 'f3', 'f4', 'f5', 'f6', 'f7', 'f8', 'f9', 'f10', 'f11', 'f12',
-  // Symbols
-  '.', ',', '/', '\\', '[', ']', ';', '\'', '`', '-', '=', 'plus'
-];
-
-/**
- * Checks if a key combination might be dangerous
- * @param keys Array of keys in the combination
- * @returns String with reason if dangerous, null if safe
- */
-function isDangerousKeyCombination(keys: string[]): string | null {
-  const normalizedKeys = keys.map(k => k.toLowerCase());
-  
-  // Check for OS-level dangerous combinations
-  if (normalizedKeys.includes('command') || normalizedKeys.includes('control')) {
-    // Control+Alt+Delete or Command+Option+Esc (Force Quit on Mac)
-    if ((normalizedKeys.includes('control') && normalizedKeys.includes('alt') && normalizedKeys.includes('delete')) ||
-        (normalizedKeys.includes('command') && normalizedKeys.includes('alt') && normalizedKeys.includes('escape'))) {
-      return 'This combination can trigger system functions';
-    }
-    
-    // Windows key combinations
-    if (normalizedKeys.includes('command') && normalizedKeys.includes('r')) {
-      return 'This combination can open the Run dialog';
-    }
-    
-    if ((normalizedKeys.includes('control') || normalizedKeys.includes('command')) && 
-        (normalizedKeys.includes('alt') || normalizedKeys.includes('option')) && 
-        normalizedKeys.includes('t')) {
-      return 'This combination can open a terminal';
-    }
-  }
-  
-  return null;
-}
+import { 
+  MAX_TEXT_LENGTH, 
+  validateKey,
+  validateKeyCombination, 
+  validateKeyHoldOperation
+} from '../../tools/validation.js';
 
 /**
  * Keysender implementation of the KeyboardAutomation interface
  */
 export class KeysenderKeyboardAutomation implements KeyboardAutomation {
   private keyboard = new Hardware().keyboard;
-
-  /**
-   * Validate a key and convert it to KeyboardButton
-   * @param key Key to validate
-   * @returns KeyboardButton
-   * @throws Error if key is invalid
-   */
-  private validateKey(key: string): KeyboardButtonType {
-    if (!key || typeof key !== 'string') {
-      throw new Error(`Invalid keyboard key: ${key}`);
-    }
-    
-    // Normalize key to lowercase for validation
-    const normalizedKey = key.toLowerCase();
-    
-    if (!VALID_KEYS.includes(normalizedKey)) {
-      throw new Error(`Invalid key: "${String(key)}". Must be one of the allowed keys.`);
-    }
-    
-    return key;
-  }
 
   typeText(input: KeyboardInput): WindowsControlResponse {
     try {
@@ -117,8 +51,8 @@ export class KeysenderKeyboardAutomation implements KeyboardAutomation {
 
   pressKey(key: string): WindowsControlResponse {
     try {
-      // Validate the key and convert to KeyboardButton
-      const keyboardKey = this.validateKey(key);
+      // Validate the key using shared validation function
+      const keyboardKey = validateKey(key);
       
       // Start the asynchronous operation and handle errors properly
       this.keyboard.sendKey(keyboardKey)
@@ -141,20 +75,8 @@ export class KeysenderKeyboardAutomation implements KeyboardAutomation {
 
   async pressKeyCombination(combination: KeyCombination): Promise<WindowsControlResponse> {
     try {
-      // Validate the key combination
-      if (!combination.keys || !Array.isArray(combination.keys) || combination.keys.length === 0) {
-        throw new Error('Key combination must contain at least one key');
-      }
-
-      if (combination.keys.length > 5) {
-        throw new Error(`Too many keys in combination: ${combination.keys.length} (max 5)`);
-      }
-
-      // Check for potentially dangerous key combinations
-      const dangerous = isDangerousKeyCombination(combination.keys);
-      if (dangerous) {
-        throw new Error(`Potentially dangerous key combination: ${combination.keys.join('+')}. ${dangerous}`);
-      }
+      // Validate the key combination using shared validation function
+      validateKeyCombination(combination);
 
       // Store original keys for the message
       const keysForMessage = [...combination.keys];
@@ -163,7 +85,7 @@ export class KeysenderKeyboardAutomation implements KeyboardAutomation {
       // Validate each key and collect press promises
       const validatedKeys: KeyboardButtonType[] = [];
       for (const key of combination.keys) {
-        const keyboardKey = this.validateKey(key);
+        const keyboardKey = validateKey(key);
         validatedKeys.push(keyboardKey);
         
         // Collect all promises to handle them properly
@@ -210,7 +132,7 @@ export class KeysenderKeyboardAutomation implements KeyboardAutomation {
         const cleanupPromises: Promise<void>[] = [];
         for (const key of combination.keys) {
           try {
-            const keyboardKey = this.validateKey(key);
+            const keyboardKey = validateKey(key);
             cleanupPromises.push(
               this.keyboard.toggleKey(keyboardKey, false)
                 .catch(err => {
@@ -237,43 +159,16 @@ export class KeysenderKeyboardAutomation implements KeyboardAutomation {
 
   async holdKey(operation: KeyHoldOperation): Promise<WindowsControlResponse> {
     try {
-      // Validate state
-      if (operation.state !== 'down' && operation.state !== 'up') {
-        throw new Error(`Invalid state: ${String(operation.state)}. Must be either 'down' or 'up'`);
-      }
-
-      // Validate duration for key down operations
-      if (operation.state === 'down') {
-        const MAX_HOLD_DURATION = 10000; // 10 seconds maximum hold time
-        const MIN_HOLD_DURATION = 10;    // 10ms minimum hold time
-        
-        if (!operation.duration) {
-          throw new Error('Duration is required for key down operations');
-        }
-        
-        if (typeof operation.duration !== 'number' || isNaN(operation.duration)) {
-          throw new Error(`Invalid duration: ${operation.duration}. Must be a number`);
-        }
-        
-        if (operation.duration < MIN_HOLD_DURATION) {
-          throw new Error(`Duration too short: ${operation.duration}ms (min ${MIN_HOLD_DURATION}ms)`);
-        }
-        
-        if (operation.duration > MAX_HOLD_DURATION) {
-          throw new Error(`Duration too long: ${operation.duration}ms (max ${MAX_HOLD_DURATION}ms)`);
-        }
-      }
-      
-      // Validate and map the key to KeyboardButton
-      const keyboardKey = this.validateKey(operation.key);
+      // Validate key hold operation using shared validation function
+      validateKeyHoldOperation(operation);
       
       // Toggle the key state (down/up)
-      await this.keyboard.toggleKey(keyboardKey, operation.state === 'down');
+      await this.keyboard.toggleKey(operation.key, operation.state === 'down');
 
       // If it's a key press (down) with duration, wait for the specified duration then release
       if (operation.state === 'down' && operation.duration) {
         await new Promise(resolve => setTimeout(resolve, operation.duration));
-        await this.keyboard.toggleKey(keyboardKey, false);
+        await this.keyboard.toggleKey(operation.key, false);
       }
 
       return {
@@ -286,8 +181,7 @@ export class KeysenderKeyboardAutomation implements KeyboardAutomation {
       // Ensure key is released in case of error during hold
       if (operation.state === 'down') {
         try {
-          const keyboardKey = this.validateKey(operation.key);
-          await this.keyboard.toggleKey(keyboardKey, false);
+          await this.keyboard.toggleKey(operation.key, false);
         } catch (releaseError) {
           console.error(`Error releasing key ${operation.key} during cleanup:`, releaseError);
           // Ignore errors during cleanup
