+++ src/tools/mouse.ts
@@ -1,75 +1,39 @@
 import { MousePosition } from '../types/common.js';
 import { WindowsControlResponse } from '../types/responses.js';
 import { createAutomationProvider } from '../providers/factory.js';
+import { 
+  validateMousePosition, 
+  validateMouseButton,
+  validateScrollAmount
+} from './validation.js';
 
 // Get the automation provider
 const provider = createAutomationProvider();
 
-// Constants for validation
-const MAX_ALLOWED_COORDINATE = 10000; // Reasonable maximum screen dimension
-
 // Define button types
 type MouseButton = 'left' | 'right' | 'middle';
-
-/**
- * Validates mouse position against screen bounds
- * @param position Position to validate
- * @returns Validated position
- * @throws Error if position is invalid or out of bounds
- */
-function validateMousePosition(position: MousePosition): MousePosition {
-  // Basic type validation
-  if (!position || typeof position !== 'object') {
-    throw new Error('Invalid mouse position: position must be an object');
-  }
-  
-  if (typeof position.x !== 'number' || typeof position.y !== 'number') {
-    throw new Error(`Invalid mouse position: x and y must be numbers, got x=${position.x}, y=${position.y}`);
-  }
-  
-  if (isNaN(position.x) || isNaN(position.y)) {
-    throw new Error(`Invalid mouse position: x and y cannot be NaN, got x=${position.x}, y=${position.y}`);
-  }
-
-  // In test environments, screen size might not be available
-  // Check for process.env or other indicators if needed
-  try {
-    // Check if position is within reasonable bounds first (this always applies)
-    if (position.x < -MAX_ALLOWED_COORDINATE || position.x > MAX_ALLOWED_COORDINATE || 
-        position.y < -MAX_ALLOWED_COORDINATE || position.y > MAX_ALLOWED_COORDINATE) {
-      throw new Error(`Position (${position.x},${position.y}) is outside reasonable bounds (-${MAX_ALLOWED_COORDINATE}, -${MAX_ALLOWED_COORDINATE})-(${MAX_ALLOWED_COORDINATE}, ${MAX_ALLOWED_COORDINATE})`);
-    }
-    
-    // Check against actual screen bounds
-    try {
-      const screenSizeResponse = provider.screen.getScreenSize();
-      if (screenSizeResponse.success && screenSizeResponse.data) {
-        const screenSize = screenSizeResponse.data as { width: number; height: number };
-        if (position.x < 0 || position.x >= screenSize.width || 
-            position.y < 0 || position.y >= screenSize.height) {
-          throw new Error(`Position (${position.x},${position.y}) is outside screen bounds (0,0)-(${screenSize.width-1},${screenSize.height-1})`);
-        }
-      }
-    } catch (screenError) {
-      console.warn('Error checking screen bounds:', screenError);
-      // Continue without screen bounds check
-    }
-  } catch (e) {
-    // In test environment, just let it through
-    if (process.env.NODE_ENV === 'test' || process.env.VITEST) {
-      console.warn('Skipping screen bounds check in test environment');
-    } else {
-      throw e;
-    }
-  }
-  
-  return position;
-}
 
 export function moveMouse(position: MousePosition): WindowsControlResponse {
   try {
     // Validate the position
     validateMousePosition(position);
+    
+    // Additional screen bounds check if not in test environment
+    if (!(process.env.NODE_ENV === 'test' || process.env.VITEST)) {
+      try {
+        const screenSizeResponse = provider.screen.getScreenSize();
+        if (screenSizeResponse.success && screenSizeResponse.data) {
+          const screenSize = screenSizeResponse.data as { width: number; height: number };
+          if (position.x < 0 || position.x >= screenSize.width || 
+              position.y < 0 || position.y >= screenSize.height) {
+            throw new Error(`Position (${position.x},${position.y}) is outside screen bounds (0,0)-(${screenSize.width-1},${screenSize.height-1})`);
+          }
+        }
+      } catch (screenError) {
+        console.warn('Error checking screen bounds:', screenError);
+        // Continue without screen bounds check
+      }
+    }
     
     return provider.mouse.moveMouse(position);
   } catch (error) {
@@ -78,24 +42,6 @@ export function moveMouse(position: MousePosition): WindowsControlResponse {
       message: `Failed to move mouse: ${error instanceof Error ? error.message : String(error)}`
     };
   }
-}
-
-/**
- * Validates mouse button
- * @param button Button to validate
- * @returns Validated button
- * @throws Error if button is invalid
- */
-function validateMouseButton(button: unknown): MouseButton {
-  if (!button || typeof button !== 'string') {
-    throw new Error(`Invalid mouse button: ${String(button)}`);
-  }
-  
-  if (!['left', 'right', 'middle'].includes(button)) {
-    throw new Error(`Invalid mouse button: ${button}. Must be 'left', 'right', or 'middle'`);
-  }
-  
-  return button as MouseButton;
 }
 
 export function clickMouse(button: MouseButton = 'left'): WindowsControlResponse {
@@ -142,15 +88,7 @@ export function getCursorPosition(): WindowsControlResponse {
 export function scrollMouse(amount: number): WindowsControlResponse {
   try {
     // Validate amount
-    if (typeof amount !== 'number' || isNaN(amount)) {
-      throw new Error(`Invalid scroll amount: ${amount}. Must be a number`);
-    }
-    
-    // Limit the maximum scroll amount
-    const MAX_SCROLL_AMOUNT = 1000;
-    if (Math.abs(amount) > MAX_SCROLL_AMOUNT) {
-      throw new Error(`Scroll amount too large: ${amount} (max ${MAX_SCROLL_AMOUNT})`);
-    }
+    validateScrollAmount(amount);
     
     return provider.mouse.scrollMouse(amount);
   } catch (error) {
